Remove deprecated back-compat things
DiscardLogger became a LogSink, so leaving that as a type was only
needed for NullLogger.

NullLogger can be completely replaced with Discard().

InfoLogger is vestigial and since this is a breaking change anyway,
let's kill it off.

Not that Logger is a type, we do not need WithCallDepth() as a
standalone anymore.