funcr: support embedding, use for example and testing logger
This allows custom loggers to provide their own additional methods
while using most of the implementation from funcr.

That funcr had to be extended to support the helper concept in
testing.T didn't feel right because fnlogger then always implemented
the CallStackHelperLogSink interface even when the helper was
nil. This gets fixed by moving the implementation of that into the
testing logger, using the new embedding mechanism of funcr.

As a drive-by fix, the testing logger no longer prints a redundant ":"
when the prefix is empty.

The logr package describes how to implement a custom With* function. A
runable example for that gets added in funcr/example_test.go.