fix incremental WithCallDepth
When WithCallDepth stored the new sink, it didn't update the cached
withCallDepth implementation. When that then was called again by
another WithCallDepth, the previously incremented call depth was
overwritten.

To ensure that users update the sink only through the new WithSink
method, the field gets unexported again. GetSink now provides
read-only access. WithSink is used instead of SetSink because the
latter would be unsafe to call when the logger instance is used
concurrently. WithSink avoids that by always creating a copy before
doing the modification.