funcr: Add options for hooks during rendering
Add new `Options` fields which are function pointers to optional hooks
to be called.

This allow users to manipulate the way that kv-pairs are logged.  For
example:
* prefix or rename builtin keys
* log WithValues() kv-pairs as a single pseudo-struct
* collect telemetry around how often each log is triggered

This adds a new type `PseudoStruct` which can be used to trigger special
logging - a kvList is treated as a struct.

This also exposes the `Caller` type, to enable hooks to recognize and
use caller information.