funcr: Prevent stack overflow on recursive structs
Add a new Option `MaxLogDepth` which  tells funcr how many levels of
nested fields (e.g. a struct that contains a struct that contains a
struct, etc.) it may log.  Every time it finds a struct, slice, array,
or map the depth is increased by one.  When the maximum is reached, the
value will be converted to a string indicating that the max depth has
been exceeded.  If this field is not specified, a default value will be
used.