add support for logging values stored in the context
Having to add a modified logger to a context when adding some values
to the context which are meant to be logged (like trace ID) has two
drawbacks:
  - the instrumentation code which adds the values to the context must
    be aware of logging
  - modifying the logger incurs a cost, whether some actual log entry
    then gets emitted or not.

A better approach is to add the values only to the context, then during logging
extract them. This is the same approach that contextual logging in Go is going
to use. The original plan was to tell a logger which keys to use to extract
values from a context. That plan was foiled by OpenTelemetry not exporting
those keys...  Instead, a function provided by OpenTelemetry must be used to
get a SpanContext, then the span ID and trace ID can be extracted from that for
logging.

Such support is possible in Logger by manipulating the key/value slice before
passing it to the LogSink. But that then needs to deal with duplicates because
the LogSink might not expect those as part of the parameters. Using WithValues
is simpler, albeit also a bit slower.

Storing new fields in Logger is simple, but slows down scenarios where those
fields are not used. A better approach will follow in the next commits.

Discard() is documented as always comparing equal. Adding a context
or context keys breaks that. It seems useful to treat such a Logger
in a special way and not modify it to preserve the promise. The same
is done for the null Logger.