add contextual logging benchmark and unit test
These benchmark tests logging of two values (like trace and span ID)
for different scenarios:
- 1% of the call chains invoke Info, the rest don't log anything
- 100% of the call chains log one entry, but it gets discarded due
  to verbosity
- 100% of the call chains log one entry which really gets emitted

Each of these is tested with and without additional calls to WithName and
WithValues.

Without special support for contextual logging, the code which adds values to
the context must also modify the logger.

The unit test reuses the setup code and checks in the benchmark functions.