add KeysAndValues as special type
This is like the existing funcr.PseudoStruct except that it
uses structs as members of the slice. The reasons for that difference are
making the API type-safe at compile time and (subjectively) nicer rendering
when a LogSink does not support the type.

Here's how zapr and klogr handle funcr.PseudoStruct:

    {"caller":"test/output.go:435","msg":"funcr","v":0,"parent":["a",1,"b",2]}
    I output.go:486] "funcr" parent=[a 1 b 2]

Here's the same for logr.KeysAndValues:

    {"caller":"test/output.go:435","msg":"keys and values","v":0,"parent":[{"Key":"a","Value":1},{"Key":"b","Value":2}]}
    I output.go:486] "keys and values" parent=[{Key:a Value:1} {Key:b Value:2}]

The tentative support for logr.KeysAndValues would lead to output like this:

    {"caller":"test/output.go:<LINE>","msg":"keys and values","v":0,"parent":{"boolsub":true,"intsub":1,"recursive":[{"Key":"sub","Value":"level2"}]}}
    I output.go:<LINE>] "keys and values" parent={ boolsub=true intsub=1 recursive={ sub="level2" } }

The implementation in klogr is recursive and handles KeysAndValues nested
inside KeysAndValues, the one in zapr is simpler and only handles it at the top
level.