cache result of context value retrieval
The underlying idea is that each contextKV instance optionally also contains a
LogSink where the values were stored.

The Logger struct can get copied, as before. Updates to it in the different
With* functions become more complicated: if there is a contextMeta instance,
then a new one must get created. The old one may still be in use by the parent
Logger and therefore must not get modified.