support slog
This enables the use of the logr.Logger API on top of a slog.Handler. The other
direction is also possible, but some information is lost. Instead, logr
implementations should better get extended for usage as slog.Handler.

Going back and forth between the two APIs is now part of logr.