slogr: add SlogSink
A LogSink may implement this additional interface if it wants to support
logging through slog better. Doing so addresses several issues:

- stack unwinding gets avoided in favor of logging the pre-recorded PC
- proper grouping of key/value pairs via WithGroup
- verbosity levels > slog.LevelInfo can be recorded
- less overhead

A LogSink implementation which wants to provide a `New*Handler() slog.Handler`
function can either implement a second type (required because the prototype of
the Enabled method conflicts) or use `logr.NewSlogHandler`. The latter is
better because then conversion back-and-forth works. The overhead for the
additional wrapper should be minimal.