move slogr into main package
This is necessary to give the context handling functions access to the slogr
functionality. Functions get renamed to match names in the main package
where needed.

The slogr package is now deprecated, but continues to be supported as part of
the logr v1 API by keeping all exported items as wrappers or aliases for the
main package.