unit tests with full SlogSink mock implementation
This turns testLogSink into an implementation of SlogSink if slog is supported,
then runs through some scenarios where values get added using different APIs.
The end result gets compared by formatting with the slog JSON handler because
that is easier than writing custom code for a deep-equal comparison of slog
attributes.

The "with-group-and-logr-values" is the case which was broken.